Import/migrate existing DEK into the new Keychain wallet
